Start trying to refactor away OrbitImpl
Okay, so we want to be able to get the initial state vector, even for orbits that are initialized using orbital elements. Given the overall architecture, it doesn't make any sense to pimpl the Orbit class. We want to guarantee that all the non-element 'auxiliary' members are initialized, so we should just do that directly in the constructor.

Anyway, as part of this, we need to move the NewtonSolver into the orbital library, which will happen next