Revert previous 'fixes', add trivial orbit test
The previous fixes might have somehow produced the right numbers, but they are physically incorrect. So add a test with a very trivial orbit, nearly circular with radius 1, velocity 1.

See that the flaw is in the true anomaly calculation is flawed. the calculation is for *cos* true anomaly, so it's missing the arccos