Correct trueAnomaly calculation
Equation used was actually cos(true anomaly) = stuff. But can't just use acos, because the quadrants problem leaves an ambiguity. So need to use atan2 to properly generate it.

Also fix, in orbit, the eccentric anomaly calculation to produce 0 rather than a divide by zero when eccentricity is 0 (circular orbits)