Inspect why calculation is still failing
Added a test in Rotation Matrices to test precisely the calculation that occurs in Orbit State, and it passes. Noticed, however, that the incoming txfm matrix from the Orbit State calculation is inverted. 

I might misunderstand something about the Matrix API and that might be throwing off my tests in a consistent way to hide the flaw here.