Tune OrbitalHIstory test for smaller test
By reducing to less than 2pi times, we don't need to deal with 'past the end' searching in the history.

By shifting the time just a moment before the boundary time, we actually get the right value. Since our search times match up precisely with the calculated nodes, we were getting the next node's result. 

On the latter point, it seems like having the interpolation algorithm in place will be useful for not having artifacts on nodes.