Add some basic Engine, Orbital .proto files
You can see in the previous commits I added, then removed protobuf as a submodule. Realized that protobuf really is its own thing that needs to be compiled and installed separately and externally to the main project. But, this does serve as a record that I'm targeting the 3.6.0.1 build of protobuf, in terms of what I'm building around.