Start a CelestialSystem class
This will be the manager for the Bodies and orbits and such. It does seem that the initial design of 'bodies' is a little off compared to what we need here.

Bodies likely will be initialised using orbital elements, masses, and parents; they won't need state-vector style constructions. Bodies will need a function to set a new parent/orbit, and keep the body persistent.