First attempt at the recursive system creation
Tests fail with map access failures to be investigated. But this really highlights the major change from a 'nested' structure where a body 'owns' its child bodies and those, in turn, own other bodies to a 'flat' structure where we have a dictionary of bodies that simply "know" who their parent is and what their orbit is around it.

So in constructing it we find the root body, then its children, then for each child, add its children and so on down the tree.