Start a separate RpcServerTest file
Not sure why, but when it's in the source file, it ends up getting doubled going into main.

Overall though, I think I'm preferring having testing be separate files from the source code. It was an interesting experiment, but because the tests are in each translation unit, running tests on any one library runs all of the tests in the entire project (at least all the ones indirectly referred to by that translation unit)