Refactor engine tests
OrbitStateTest had some tests of helper functions that we now can discard. Maybe this is actually a good overall pattern? Maybe when we're building the scaffolding of the classes, having the tests in source helps us poke at that scaffolding. Or perhaps the helper stuff needs to be explicitly in its own source to be tested.