Clean up unused functions in Orbit
Here's the thing: The helper functions in Orbit are about taking state vector to orbital elements, and the helper functions in Orbit State are about the reverse operation. They don't actually overlap much. If there is need eventually to come up with a common 'helper functions' file, we'll make that refactoring then.