Add an addBodyDialog
Plan here is that the dialog has the orbit info greyed out unless there's at least one parent body in the system. That way the first body entered must be the root parent body. Then, if there are parents available, add them to the parent selector combo box and enable the orbit info.

Not doing any data sanitization yet, and could add in a lot more options for a user here.